Version 0.9

+++ service/app/users/router.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Response, Request
+from fastapi import APIRouter, Response, Request, HTTPException
 from pydantic import EmailStr
 
 from users.validation import SUser_registration, SUser_authentication
@@ -24,6 +24,7 @@ def user_register(data: SUser_registration, response: Response) -> dict:
     # Заменяем пароль на хэшированный
     data = data.model_copy(update={"password": hash_password(data.password)})
 
+    # Если пользователь не активен, то обновляем уже существующие данные
     if found:
         UsersDAO.update_user(
             email=data.email, 
@@ -45,16 +46,31 @@ def user_register(data: SUser_registration, response: Response) -> dict:
 
 
 @router.post("/logout/", summary="Выход из профиля")
-def user_logout(response: Response) -> dict:
+def user_logout(response: Response, request: Request) -> dict:
     """Разлогинивает пользователя."""
 
+    token = request.cookies.get("users_access_token")
+    if token is None:
+        raise HTTPException(
+            status_code=401, 
+            detail="Пользователь не авторизован"
+        )
+    
     response.delete_cookie(key="users_access_token")
     return {"message": "Пользователь успешно разлогинен."}
 
 
 @router.post("/login/", summary="Аутентификация пользователя")
-def user_login(data: SUser_authentication, response: Response) -> dict:
+def user_login(
+    data: SUser_authentication,
+    response: Response,
+    request: Request
+) -> dict:
     """Входит в аккаунт пользователя."""
+
+    token = request.cookies.get("users_access_token")
+    if token is not None:
+        return {"message": "Пользователь уже авторизован."}
     
     token = create_access_token(email=data.email)
     response.set_cookie(key="users_access_token", value=token, httponly=True)
@@ -64,14 +80,20 @@ def user_login(data: SUser_authentication, response: Response) -> dict:
 
 @router.post("/delete/", summary="Удаление пользователя")
 def user_delete(request: Request, response: Response) -> dict:
-    """Удаляет аккаунт пользователя из базы данных"""
+    """Удаляет аккаунт пользователя из базы данных."""
 
     token = request.cookies.get("users_access_token")
+    if token is None: 
+        raise HTTPException(
+            status_code=401, 
+            detail="Пользователь не авторизован"
+        )
+
     user_email = decode_access_token(token)
     UsersDAO.delete_user(user_email)
 
     response.delete_cookie(key="users_access_token")
-    return {"message": "Пользователь успешно удален"}
+    return {"message": "Пользователь успешно удален."}
 
 
 @router.post("/update/", summary="Обновление данные пользователя")
@@ -79,19 +101,27 @@ def user_update(data: SUser_update_data, request: Request) -> dict:
     """Обновляет данные пользователя."""
 
     token = request.cookies.get("users_access_token")
+    if token is None:
+        raise HTTPException(
+            status_code=401, 
+            detail="Пользователь не авторизован"
+        )
+    
     user_email = decode_access_token(token)
 
+    # Удаляем не введённые поля
     data = data.model_dump()
     for key in list(data.keys()):
         if data[key] is None:
             del data[key]
 
+    # Хэшируем новый пароль, если он есть
     if data.get("password") is not None:
         data["password"] = hash_password(data["password"])
 
     UsersDAO.update_user(email=user_email, **data)
 
-    return {"message": "Данные успешно изменены"}
+    return {"message": "Данные успешно изменены."}
 
 
 @router.get("/data/", summary="Получение данных о пользователе")
@@ -100,28 +130,31 @@ def user_data(email: EmailStr, request: Request) -> dict:
     """Показывает данные о пользователе."""
 
     user = UsersDAO.find_user(email=email)
+    if user == False:
+        return {"message": "Такого пользователя нет"}
+    
     return user.get_dict()
 
 
-@router.get("/rules", summary="Показ правил для админа")
+@router.get("/rules", summary="Показ правил админа")
 @require_role(role="admin")
 def get_admin_rules(request: Request) -> dict:
-    """Показывает правила для админа"""
+    """Показывает правила админа."""
 
     return AdminRules.rules
 
 
-@router.post("/rules_add", summary="Добавление правила в правила админа")
+@router.post("/rules_add", summary="Добавление правила в список правил админа")
 @require_role(role="admin")
 def add_admin_rules(new_rules: str, request: Request) -> dict:
-    """Добавляет правило в правила админа"""
+    """Добавляет правило в список правил админа."""
 
     return AdminRules.add_rules(new_rule=new_rules)
 
 
-@router.post("/rules_del", summary="Удаление правила из правил админа")
+@router.post("/rules_del", summary="Удаление правила из списка правил админа")
 @require_role(role="admin")
 def del_admin_rules(number_rule: int, request: Request) -> dict:
-    """Удаляет правило из правил админа"""
+    """Удаляет правило из списка правил админа."""
 
     return AdminRules.del_rules(number_rule=number_rule)
